+++ module/woocommerce/config.go
@@ -1,6 +1,0 @@
-package woocommerce
-
-import wooruntime "mannaiah/module/woocommerce/runtime"
-
-// Config defines WooCommerce integration configuration values.
-type Config = wooruntime.Config